# PERF: improve IntegerArray fast constructor