--- a/pandas/core/arrays/integer.py
+++ b/pandas/core/arrays/integer.py
@@ -343,12 +343,12 @@ class IntegerArray(BaseMaskedArray):
         return _dtypes[str(self._data.dtype)]
 
     def __init__(self, values: np.ndarray, mask: np.ndarray, copy: bool = False):
-        if not (isinstance(values, np.ndarray) and is_integer_dtype(values.dtype)):
+        if not (isinstance(values, np.ndarray) and values.dtype.kind in ["i", "u"]):
             raise TypeError(
                 "values should be integer numpy array. Use "
                 "the 'integer_array' function instead"
             )
-        if not (isinstance(mask, np.ndarray) and is_bool_dtype(mask.dtype)):
+        if not (isinstance(mask, np.ndarray) and mask.dtype.kind == "b"):
             raise TypeError(
                 "mask should be boolean numpy array. Use "
                 "the 'integer_array' function instead"