@jorisvandenbossche 2020-04-07
Eg `is_bool_dtype` does:

- check the value (if it is an array, get the dtype, if it is a string, convert to a dtype object, etc)
- check if it is a CategoricalDtype, and in that case check the dtype of the categories
- check whether the passed object was an Index, in which case object dtype can be inferred to be boolean
- check if it is an extension dtype, in which case we check the `_is_boolean` property
- finally check `issubclass(dtype.type, np.bool_)` (which is also still slightly slower than `dtype == np.bool_` or `dtype.kind == "b"`, but I am not fully sure there are practical differences)

So we need to keep all those checks I think, but we might be able to start that function with a short-cut check for the special case of a numpy dtype (see my third bullet point above)