@simonjayhawkins 2020-04-07
> For optimal performance, having a way to don't do any validation at all (so also don't check the passed values are ndarray, or their dtypes) would actually even be better than what this PR does. But I also like the clean constructor we have now (eg no separate private constructor)

IIUC constructing integer array from the IntegerArray constructor is not part of the public api, so just removing the validation is also an option?