@TomAugspurger 2020-04-07
> IIUC constructing integer array from the IntegerArray constructor is not part of the public api, so just removing the validation is also an option?

This is a bit hard, since you can do things like `type(pd.array([1, 2, 3]))(...)`, which uses just public APIs to call the IntegerArray constructor. It's kinda hard to make a `__init__` private.