@jorisvandenbossche 2020-04-07
You can't make the init private, but that doesn't mean you need to accept all kinds of values *or* do validation of the input in the init. It can be the documented behaviour (now, I suppose in the public init we do want *some* basic validation, to avoid people shooting themselves in the foot. But it doesn't need to be, or it certainly can be the bare minimum)