@jorisvandenbossche 2020-04-07
> I'd be happy with deprecating some of those behaviors, but that's a separate issue :)

Do you think any of those items I listed can be deprecated? 
Passing an array instead of dtype, yes (which also means the object dtype Index can't be checked anymore), but for the rest I think we want to keep those behaviours?