@jbrockmendel 2020-04-07
> I'd be happy with deprecating some of those behaviors, but that's a separate issue :)

Stumbled across this comment moments after opening #33368