@jbrockmendel 2020-04-07
LGTM.  The dtype.kind checks are way more performant than the is_foo_dtype, especially when checking multiple possibilities like `["i", "u"]`