@jorisvandenbossche 2020-04-10
Are we ok with moving forward with this? 
Even if we make the `is_.._dtype` checks faster (xref the other PRs/discussions), they probably can never beat `dtype.kind in ["i", "u"]` in case you already checked `dtype` to be a numpy dtype (eg `is_integer_dtype` will also need to check for extension integer dtype).

I am personally fine with having the more specialized check here in the constructor.