@TomAugspurger 2020-04-10
I'm fine with the specialized / high-performance checks here.