@jreback 2020-04-10
thanks @jorisvandenbossche yep a specialized check a-ok with me